+++ pubspec.yaml
@@ -24,6 +24,9 @@ dependencies:
   flutter:
     sdk: flutter
 
+  flutter_login_signup:
+    git:
+      url: https://github.com/jpvazque/flutter_login_signup.git
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
